# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/organizations/{organizationId}/sites/{siteId}/pages/{pageId}/blocks": {
      "get": {
        "description": "En orden de `position`, con la configuración de su versión vigente. Un bloque que no se puede renderizar viene con `degraded` en vez de romper la respuesta: el render lo omite y el constructor puede avisar. Solo pide membresía activa.",
        "operationId": "BlocksController_list",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Organización dueña del recurso. Resuelve el tenant antes de tocar nada (ADR-002).",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "siteId",
            "required": true,
            "in": "path",
            "description": "Sitio dueño de la página.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "pageId",
            "required": true,
            "in": "path",
            "description": "Página dueña de los bloques.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "X-Requested-With",
            "in": "header",
            "required": false,
            "description": "Obligatoria con valor `impulza-one` en toda petición que modifica estado. Ver CsrfGuard.",
            "schema": {
              "type": "string",
              "enum": [
                "impulza-one"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bloques de la página, ordenados.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      "pageId": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      "type": {
                        "type": "string"
                      },
                      "position": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "configSchemaVersion": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "visible": {
                        "type": "boolean"
                      },
                      "scheduledStart": {
                        "nullable": true,
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                      },
                      "scheduledEnd": {
                        "nullable": true,
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                      },
                      "config": {},
                      "degraded": {
                        "nullable": true,
                        "type": "string",
                        "enum": [
                          "unknown_type",
                          "future_version",
                          "invalid_config"
                        ]
                      }
                    },
                    "required": [
                      "id",
                      "pageId",
                      "type",
                      "position",
                      "configSchemaVersion",
                      "visible",
                      "scheduledStart",
                      "scheduledEnd",
                      "config",
                      "degraded"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida (detalle en `issues`)."
          },
          "401": {
            "description": "Sin sesión válida."
          },
          "403": {
            "description": "Sin membresía activa en la organización, sin el permiso requerido, o petición sin la cabecera anti-CSRF."
          },
          "404": {
            "description": "Página no encontrada, o de otra organización (ADR-002)."
          }
        },
        "security": [
          {
            "sesion": []
          }
        ],
        "summary": "Listar los bloques de la página",
        "tags": [
          "blocks"
        ]
      },
      "post": {
        "description": "Requiere `page.manage`. `type` sale del catálogo cerrado y `config` se valida contra el esquema **de ese tipo** y se sanitiza en el servidor — es el único camino por el que una configuración llega a la base, que es lo que convierte «bloques tipados, no HTML arbitrario» en una garantía y no en una intención.",
        "operationId": "BlocksController_create",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Organización dueña del recurso. Resuelve el tenant antes de tocar nada (ADR-002).",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "siteId",
            "required": true,
            "in": "path",
            "description": "Sitio dueño de la página.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "pageId",
            "required": true,
            "in": "path",
            "description": "Página donde se agrega el bloque.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "X-Requested-With",
            "in": "header",
            "required": false,
            "description": "Obligatoria con valor `impulza-one` en toda petición que modifica estado. Ver CsrfGuard.",
            "schema": {
              "type": "string",
              "enum": [
                "impulza-one"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "profile",
                      "hero",
                      "text",
                      "link",
                      "social",
                      "image",
                      "gallery",
                      "video",
                      "whatsapp",
                      "contact_actions",
                      "contact_form",
                      "service",
                      "divider",
                      "faq",
                      "testimonials"
                    ]
                  },
                  "config": {},
                  "visible": {
                    "type": "boolean"
                  },
                  "scheduledStart": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                  },
                  "scheduledEnd": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                  }
                },
                "required": [
                  "type",
                  "config"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bloque creado, al final de la página.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "pageId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "type": {
                      "type": "string"
                    },
                    "position": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "configSchemaVersion": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "visible": {
                      "type": "boolean"
                    },
                    "scheduledStart": {
                      "nullable": true,
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    },
                    "scheduledEnd": {
                      "nullable": true,
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    },
                    "config": {},
                    "degraded": {
                      "nullable": true,
                      "type": "string",
                      "enum": [
                        "unknown_type",
                        "future_version",
                        "invalid_config"
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "pageId",
                    "type",
                    "position",
                    "configSchemaVersion",
                    "visible",
                    "scheduledStart",
                    "scheduledEnd",
                    "config",
                    "degraded"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Tipo de bloque fuera del catálogo, o `scheduledStart` posterior o igual a `scheduledEnd`."
          },
          "401": {
            "description": "Sin sesión válida."
          },
          "403": {
            "description": "Sin membresía activa en la organización, sin el permiso requerido, o petición sin la cabecera anti-CSRF."
          },
          "404": {
            "description": "Página no encontrada, o de otra organización (ADR-002)."
          },
          "422": {
            "description": "La configuración no cumple el esquema de ese tipo de bloque. `issues` indica el campo exacto."
          }
        },
        "security": [
          {
            "sesion": []
          }
        ],
        "summary": "Agregar un bloque a la página",
        "tags": [
          "blocks"
        ]
      }
    },
    "/api/v1/organizations/{organizationId}/sites/{siteId}/pages/{pageId}/blocks/reorder": {
      "put": {
        "description": "Requiere `page.manage`. Mismo criterio que las páginas: se manda el orden completo, con exactamente todos los bloques de la página y sin repetidos.",
        "operationId": "BlocksController_reorder",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Organización dueña del recurso. Resuelve el tenant antes de tocar nada (ADR-002).",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "siteId",
            "required": true,
            "in": "path",
            "description": "Sitio dueño de la página.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "pageId",
            "required": true,
            "in": "path",
            "description": "Página cuyos bloques se reordenan.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "X-Requested-With",
            "in": "header",
            "required": false,
            "description": "Obligatoria con valor `impulza-one` en toda petición que modifica estado. Ver CsrfGuard.",
            "schema": {
              "type": "string",
              "enum": [
                "impulza-one"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "blockIds": {
                    "minItems": 1,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    }
                  }
                },
                "required": [
                  "blockIds"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bloques en el nuevo orden.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      "pageId": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      "type": {
                        "type": "string"
                      },
                      "position": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "configSchemaVersion": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "visible": {
                        "type": "boolean"
                      },
                      "scheduledStart": {
                        "nullable": true,
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                      },
                      "scheduledEnd": {
                        "nullable": true,
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                      },
                      "config": {},
                      "degraded": {
                        "nullable": true,
                        "type": "string",
                        "enum": [
                          "unknown_type",
                          "future_version",
                          "invalid_config"
                        ]
                      }
                    },
                    "required": [
                      "id",
                      "pageId",
                      "type",
                      "position",
                      "configSchemaVersion",
                      "visible",
                      "scheduledStart",
                      "scheduledEnd",
                      "config",
                      "degraded"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "400": {
            "description": "La lista tiene identificadores repetidos, o no incluye exactamente todos los bloques de la página."
          },
          "401": {
            "description": "Sin sesión válida."
          },
          "403": {
            "description": "Sin membresía activa en la organización, sin el permiso requerido, o petición sin la cabecera anti-CSRF."
          },
          "404": {
            "description": "Página no encontrada, o de otra organización (ADR-002)."
          }
        },
        "security": [
          {
            "sesion": []
          }
        ],
        "summary": "Reordenar los bloques de la página",
        "tags": [
          "blocks"
        ]
      }
    },
...
  },
  "info": {
    "title": "Impulza One API",
    "description": "API REST de Impulza One. Todo cuelga de `/api/v1` salvo `/health`, que es un endpoint de\ninfraestructura sin sesión ni CSRF.\n\n**Autenticación**: cookie de sesión `HttpOnly` (`impulza_session`), emitida por\n`POST /auth/login`. No hay tokens en el cuerpo ni en cabeceras de autorización.\n\n**CSRF**: toda petición que modifica estado (POST, PATCH, PUT, DELETE) debe enviar la\ncabecera `X-Requested-With: impulza-one`. Un formulario o una imagen entre sitios no puede\nfijar cabeceras propias, y una petición `fetch` con cabecera custom dispara preflight CORS,\nbloqueado salvo origen autorizado. Sin ella la respuesta es 403.\n\n**Multi-tenant**: los recursos de negocio cuelgan de `/organizations/:organizationId/…`.\nUn recurso de otra organización responde **404 y no 403**, para no confirmar que existe\n(ADR-002).\n\n**Errores de validación**: 400 cuando el cuerpo no cumple el esquema de entrada, 422 cuando\ncumple la forma pero viola una regla de dominio (contraste insuficiente en un tema,\nconfiguración inválida para un tipo de bloque). Ambos traen `issues` con el campo exacto.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "sesion": {
        "type": "apiKey",
        "in": "cookie",
        "name": "impulza_session"
      }
    },
...
  }
}